requirements.txt
Flask==2.2.5
python-dotenv==1.0.1
spacy==3.7.1
SpeechRecognition==3.8.1
pyttsx3==2.90
wolframalpha==5.0.0
requests==2.31.0
beautifulsoup4==4.12.2
APScheduler==3.10.4
pyaudio==0.2.14     # may fail to build on some systems - see notes below
